+++ rye/src/pyproject.rs
@@ -5,6 +5,7 @@ use std::collections::{HashMap, HashSet};
 use std::env;
 use std::env::consts::{ARCH, OS};
 use std::ffi::OsStr;
+use std::ffi::OsString;
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::process::Command;
@@ -14,7 +15,7 @@ use std::sync::Arc;
 use anyhow::{anyhow, bail, Context, Error};
 use globset::Glob;
 use once_cell::sync::Lazy;
-use pep440_rs::{Operator, VersionSpecifiers};
+use pep440_rs::{Operator, Version, VersionSpecifiers};
 use pep508_rs::Requirement;
 use regex::Regex;
 use serde::Serialize;
@@ -24,7 +25,7 @@ use url::Url;
 use crate::config::Config;
 use crate::consts::VENV_BIN;
 use crate::platform::{get_python_version_request_from_pyenv_pin, list_known_toolchains};
-use crate::sources::{get_download_url, PythonVersion, PythonVersionRequest};
+use crate::sources::{get_download_url, matches_version, PythonVersion, PythonVersionRequest};
 use crate::sync::VenvMarker;
 use crate::utils::{
     expand_env_vars, format_requirement, get_short_executable_name, is_executable,
@@ -348,8 +349,7 @@ impl Workspace {
 
     /// Checks if a project is a member of the declared workspace.
     pub fn is_member(&self, path: &Path) -> bool {
-        let canonicalized = self.root.join(path);
-        if let Ok(relative) = path.strip_prefix(canonicalized) {
+        if let Ok(relative) = path.strip_prefix(&self.root) {
             if relative == Path::new("") || self.members.is_empty() {
                 true
             } else {
@@ -419,14 +419,14 @@ impl Workspace {
     /// That is the Python version that appears as lower bound in the
     /// pyproject toml.
     pub fn target_python_version(&self) -> Option<PythonVersionRequest> {
-        resolve_target_python_version(&self.doc, &self.venv_path())
+        resolve_target_python_version(&self.doc, &self.root, &self.venv_path())
     }
 
     /// Returns the project's intended venv python version.
     ///
     /// This is the python version that should be used for virtualenvs.
     pub fn venv_python_version(&self) -> Result<PythonVersion, Error> {
-        resolve_intended_venv_python_version(&self.doc)
+        resolve_intended_venv_python_version(&self.doc, &self.root)
     }
 
     /// Returns a list of index URLs that should be considered.
@@ -445,20 +445,44 @@ fn skip_recurse_into(name: &OsStr) -> bool {
     return name == OsStr::new(".venv") || name == OsStr::new(".git");
 }
 
+/// Could not auto-discover any pyproject
+#[derive(Debug, Clone)]
+pub struct DiscoveryUnsuccessful;
+
+impl std::error::Error for DiscoveryUnsuccessful {}
+
+impl fmt::Display for DiscoveryUnsuccessful {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "did not find pyproject.toml")
+    }
+}
+
 /// Helps working with pyproject.toml files
 #[derive(Debug)]
 pub struct PyProject {
     root: PathBuf,
+    basename: OsString,
     workspace: Option<Arc<Workspace>>,
     doc: Document,
 }
 
 impl PyProject {
+    /// Load a pyproject toml if explicitly given, else discover from current directory
+    ///
+    /// Used for command line arguments.
+    pub fn load_or_discover(arg: Option<&Path>) -> Result<PyProject, Error> {
+        match arg {
+            // canonicalize because it comes from a command line argument
+            Some(path) => Self::load(&path.canonicalize()?),
+            None => Self::discover(),
+        }
+    }
+
     /// Discovers and loads a pyproject toml.
     pub fn discover() -> Result<PyProject, Error> {
         let pyproject_toml = match find_project_root() {
             Some(root) => root.join("pyproject.toml"),
-            None => bail!("did not find pyproject.toml"),
+            None => return Err(Error::from(DiscoveryUnsuccessful)),
         };
         Self::load(&pyproject_toml)
     }
@@ -466,7 +490,8 @@ impl PyProject {
     /// Loads a pyproject toml.
     pub fn load(filename: &Path) -> Result<PyProject, Error> {
         let root = filename.parent().unwrap_or(Path::new("."));
-        let doc = fs::read_to_string(filename)?
+        let doc = fs::read_to_string(filename)
+            .with_context(|| format!("failed to read pyproject.toml from {}", &filename.display()))?
             .parse::<Document>()
             .with_context(|| {
                 format!(
@@ -490,8 +515,14 @@ impl PyProject {
             }
         }
 
+        let basename = match filename.file_name() {
+            Some(name) => name.to_os_string(),
+            None => bail!("project {} has no file name", root.display()),
+        };
+
         Ok(PyProject {
             root: root.to_owned(),
+            basename,
             workspace,
             doc,
         })
@@ -519,8 +550,14 @@ impl PyProject {
             return Ok(None);
         }
 
+        let basename = match filename.file_name() {
+            Some(name) => name.to_os_string(),
+            None => bail!("project {} has no file name", root.display()),
+        };
+
         Ok(Some(PyProject {
             root: root.to_owned(),
+            basename,
             workspace: Some(workspace),
             doc,
         }))
@@ -556,7 +593,7 @@ impl PyProject {
 
     /// Returns the path to the toml file.
     pub fn toml_path(&self) -> Cow<'_, Path> {
-        Cow::Owned(self.root.join("pyproject.toml"))
+        Cow::Owned(self.root.join(&self.basename))
     }
 
     /// Returns the location of the virtualenv.
@@ -577,7 +614,7 @@ impl PyProject {
         if let Some(workspace) = self.workspace() {
             workspace.target_python_version()
         } else {
-            resolve_target_python_version(&self.doc, &self.venv_path())
+            resolve_target_python_version(&self.doc, &self.root, &self.venv_path())
         }
     }
 
@@ -588,7 +625,7 @@ impl PyProject {
         if let Some(workspace) = self.workspace() {
             workspace.venv_python_version()
         } else {
-            resolve_intended_venv_python_version(&self.doc)
+            resolve_intended_venv_python_version(&self.doc, &self.root)
         }
     }
 
@@ -607,6 +644,16 @@ impl PyProject {
         project["requires-python"] = Item::Value(Value::String(Formatted::new(marker)));
     }
 
+    /// Set the project version.
+    pub fn set_version(&mut self, version: &Version) {
+        let project = self
+            .doc
+            .entry("project")
+            .or_insert(Item::Table(Table::new()));
+
+        project["version"] = Item::Value(Value::String(Formatted::new(version.to_string())));
+    }
+
     /// Returns the project name.
     pub fn name(&self) -> Option<&str> {
         self.doc
@@ -620,6 +667,28 @@ impl PyProject {
         self.name()
             .map(normalize_package_name)
             .ok_or_else(|| anyhow!("project from '{}' has no name", self.root_path().display()))
+    }
+
+    /// Returns the version.
+    pub fn version(&mut self) -> Result<Version, Error> {
+        let version = self
+            .doc
+            .get("project")
+            .and_then(|x| x.get("version"))
+            .and_then(|x| x.as_str());
+
+        match version {
+            Some(version) => {
+                Version::from_str(version).map_err(|msg| anyhow!("invalid version: {}", msg))
+            }
+            None => {
+                let version = Version::from_str("0.1.0").unwrap();
+                self.set_version(&version);
+                self.save()?;
+
+                Ok(version)
+            }
+        }
     }
 
     /// Returns the build backend.
@@ -879,7 +948,7 @@ pub fn latest_available_python_version(
         available
             .into_iter()
             .filter_map(|(ver, _)| {
-                if Some(&ver.kind as &str) == requested_version.kind.as_deref() {
+                if matches_version(requested_version, &ver) {
                     Some(ver)
                 } else {
                     None
@@ -898,15 +967,22 @@ pub fn latest_available_python_version(
     all.into_iter().rev().next()
 }
 
-fn resolve_target_python_version(doc: &Document, venv_path: &Path) -> Option<PythonVersionRequest> {
+fn resolve_target_python_version(
+    doc: &Document,
+    root: &Path,
+    venv_path: &Path,
+) -> Option<PythonVersionRequest> {
     resolve_lower_bound_python_version(doc)
         .or_else(|| get_current_venv_python_version(venv_path).map(Into::into))
-        .or_else(|| get_python_version_request_from_pyenv_pin().map(Into::into))
+        .or_else(|| get_python_version_request_from_pyenv_pin(root).map(Into::into))
         .or_else(|| Config::current().default_toolchain().ok())
 }
 
-fn resolve_intended_venv_python_version(doc: &Document) -> Result<PythonVersion, Error> {
-    let requested_version = get_python_version_request_from_pyenv_pin()
+fn resolve_intended_venv_python_version(
+    doc: &Document,
+    root: &Path,
+) -> Result<PythonVersion, Error> {
+    let requested_version = get_python_version_request_from_pyenv_pin(root)
         .or_else(|| resolve_lower_bound_python_version(doc))
         .or_else(|| Config::current().default_toolchain().ok())
         .ok_or_else(|| {